+++ PatchcordTests/Generics/ConnectionMock.swift
