+++ PatchcordTests/Generics/NDT7MeasurementStubs.swift
