+++ PatchcordTests/Generics/NDT7TestStub.swift
